Scorecard: update ui for indicator list

+++ db/migrate/20201022064041_create_voting_indicators.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 class CreateVotingIndicators < ActiveRecord::Migration[6.0]
   def change
     create_table :voting_indicators do |t|
       t.integer :indicatorable_id
       t.string  :indicatorable_type
       t.string  :scorecard_uuid
-      t.float   :median
+      t.integer :median
       t.text    :strength
       t.text    :weakness
       t.text    :improvement
